Moved inline functions into new directory
Refactored inline_functions into a new directory. In here we can break
up the various query classes into separate files. This will assist
commits and PRs further down the line as we can work on multiple
queries at a time without incurring massive file lengths or massive
merge conflicts.
