+++ ansys/mapdl/core/inline_functions/__init__.py
@@ -1,0 +1,2 @@
+from .inline_functions import Query
+from .core import SelectionStatus
