+++ ansys/mapdl/core/inline_functions/component_queries.py
@@ -1,67 +1,4 @@
-import warnings
-from enum import IntEnum
-
-
-class SelectionStatus(IntEnum):
-    """Enumeration class for selection status information.
-
-    This class is used with methods on the ``Query`` class and has the
-    following options.
-
-    UNSELECTED = -1
-    UNDEFINED = 0
-    SELECTED = 1
-
-    Examples
-    --------
-    The following example is taken from ``Query.nsel`` and demonstrates
-    how `SelectionSatus` appears in PyMAPDL.
-
-    Here we create a single node and interrogate its selection
-    status.
-    
-    >>> from ansys.mapdl.core import launch_mapdl
-    >>> from ansys.mapdl.core.inline_functions import Query
-    >>> mapdl = launch_mapdl()
-    >>> mapdl.prep7()
-    >>> n1 = mapdl.n(1, 0, 0, 0)
-    >>> n1
-    1
-
-    We can use ``Query.nsel`` to interrogate the selection status
-    of the node. The response is an ``enum.IntEnum`` object. If
-    you query a node that does not exist, it will return a status
-    ``SelectionStatus.UNDEFINED``.
-
-    >>> q = Query(mapdl)
-    >>> q.nsel(n1)
-    <SelectionStatus.SELECTED: 1>
-    >>> mapdl.nsel('NONE')
-    >>> q.nsel(n1)
-    <SelectionStatus.UNSELECTED: -1>
-    >>> q.nsel(0)
-    <SelectionStatus.UNDEFINED: 0>
-    """
-    UNSELECTED = -1
-    UNDEFINED = 0
-    SELECTED = 1
-
-
-class _ParameterParsing:
-    def _parse_parameter_integer_response(self, response) -> int:
-        return int(self._parse_parameter_float_response(response))
-
-    @staticmethod
-    def _parse_parameter_float_response(response) -> float:
-        if 'PARAMETER' not in response or '=' not in response:
-            raise TypeError(f'Parameter response not recognised: '
-                            f'"{response}"')
-        parts = response.rsplit('=', 1)
-        if 'WARNING' in parts[0]:
-            warnings.warn(parts[0])
-
-        number = parts[1].strip()
-        return float(number)
+from .core import _ParameterParsing
 
 
 class _ComponentQueries(_ParameterParsing):
@@ -784,492 +721,3 @@ class _DisplacementComponentQueries(_ParameterParsing):
         """
         response = self._mapdl.run(f'_=UZ({n})')
         return self._parse_parameter_float_response(response)
-
-
-class _SelectionStatusQueries(_ParameterParsing):
-    _mapdl = None
-
-    def nsel(self, n: int) -> SelectionStatus:
-        """Returns selection status of a node.
-
-        Returns a ``SelectionStatus`` object with values:
-
-        1  - SELECTED
-        0  - UNDEFINED
-        -1 - UNSELECTED
-
-        Parameters
-        ----------
-        n : int
-            Node number
-
-        Returns
-        -------
-        mapdl.ansys.core.inline_functions.SelectionStatus
-            Status of node
-
-        Examples
-        --------
-        Here we create a single node and interrogate its selection
-        status.
-
-        >>> from ansys.mapdl.core import launch_mapdl
-        >>> from ansys.mapdl.core.inline_functions import Query
-        >>> mapdl = launch_mapdl()
-        >>> mapdl.prep7()
-        >>> n1 = mapdl.n(1, 0, 0, 0)
-        >>> n1
-        1
-
-        We can use ``Query.nsel`` to interrogate the selection status
-        of the node. The response is an ``enum.IntEnum`` object. If
-        you query a node that does not exist, it will return a status
-        ``SelectionStatus.UNDEFINED``.
-
-        >>> q = Query(mapdl)
-        >>> q.nsel(n1)
-        <SelectionStatus.SELECTED: 1>
-        >>> mapdl.nsel('NONE')
-        >>> q.nsel(n1)
-        <SelectionStatus.UNSELECTED: -1>
-        >>> q.nsel(0)
-        <SelectionStatus.UNDEFINED: 0>
-        """
-        response = self._mapdl.run(f'_=NSEL({n})')
-        integer = self._parse_parameter_integer_response(response)
-        return SelectionStatus(integer)
-
-    def ksel(self, k: int) -> SelectionStatus:
-        """Returns selection status of a keypoint.
-
-        Returns a ``SelectionStatus`` object with values:
-
-        1  - SELECTED
-        0  - UNDEFINED
-        -1 - UNSELECTED
-
-        Parameters
-        ----------
-        k : int
-            Keypoint number
-
-        Returns
-        -------
-        mapdl.ansys.core.inline_functions.SelectionStatus
-            Status of keypoint
-
-        Examples
-        --------
-        Here we create a single keypoint and interrogate its selection
-        status.
-
-        >>> from ansys.mapdl.core import launch_mapdl
-        >>> from ansys.mapdl.core.inline_functions import Query
-        >>> mapdl = launch_mapdl()
-        >>> mapdl.prep7()
-        >>> k1 = mapdl.k(1, 0, 0, 0)
-        >>> k1
-        1
-
-        We can use ``Query.ksel`` to interrogate the selection status
-        of the node. The response is an ``enum.IntEnum`` object. If
-        you query a node that does not exist, it will return a status
-        ``SelectionStatus.UNDEFINED``.
-
-        >>> q = Query(mapdl)
-        >>> q.ksel(k1)
-        <SelectionStatus.SELECTED: 1>
-        >>> mapdl.ksel('NONE')
-        >>> q.ksel(k1)
-        <SelectionStatus.UNSELECTED: -1>
-        >>> q.ksel(0)
-        <SelectionStatus.UNDEFINED: 0>
-        """
-        response = self._mapdl.run(f'_=KSEL({k})')
-        integer = self._parse_parameter_integer_response(response)
-        return SelectionStatus(integer)
-
-    def lsel(self, n: int) -> SelectionStatus:
-        """Returns selection status of a line.
-
-        Returns a ``SelectionStatus`` object with values:
-
-        1  - SELECTED
-        0  - UNDEFINED
-        -1 - UNSELECTED
-
-        Parameters
-        ----------
-        n : int
-            Line number
-
-        Returns
-        -------
-        mapdl.ansys.core.inline_functions.SelectionStatus
-            Status of line
-
-        Examples
-        --------
-        Here we create a single line and interrogate its selection
-        status.
-
-        >>> from ansys.mapdl.core import launch_mapdl
-        >>> from ansys.mapdl.core.inline_functions import Query
-        >>> mapdl = launch_mapdl()
-        >>> mapdl.prep7()
-        >>> k1 = mapdl.k(1, 0, 0, 0)
-        >>> k2 = mapdl.k(2, 1, 1, 1)
-        >>> L1 = mapdl.l(k1, k2)
-        >>> L1
-        1
-
-        We can use ``Query.lsel`` to interrogate the selection status
-        of the line. The response is an ``enum.IntEnum`` object. If
-        you query a line that does not exist, it will return a status
-        ``SelectionStatus.UNDEFINED``.
-
-        >>> q = Query(mapdl)
-        >>> q.lsel(L1)
-        <SelectionStatus.SELECTED: 1>
-        >>> mapdl.lsel('NONE')
-        >>> q.lsel(L1)
-        <SelectionStatus.UNSELECTED: -1>
-        >>> q.lsel(0)
-        <SelectionStatus.UNDEFINED: 0>
-        """
-        response = self._mapdl.run(f'_=LSEL({n})')
-        integer = self._parse_parameter_integer_response(response)
-        return SelectionStatus(integer)
-
-    def asel(self, a: int) -> SelectionStatus:
-        """Returns selection status of an area.
-
-        Returns a ``SelectionStatus`` object with values:
-
-        1  - SELECTED
-        0  - UNDEFINED
-        -1 - UNSELECTED
-
-        Parameters
-        ----------
-        a : int
-            Area number
-
-        Returns
-        -------
-        mapdl.ansys.core.inline_functions.SelectionStatus
-            Selection status of the area.
-
-        Examples
-        --------
-        Here we create a single area and interrogate its selection
-        status.
-
-        >>> from ansys.mapdl.core import launch_mapdl
-        >>> from ansys.mapdl.core.inline_functions import Query
-        >>> mapdl = launch_mapdl()
-        >>> mapdl.prep7()
-        >>> k1 = mapdl.k(1, 0, 0, 0)
-        >>> k2 = mapdl.k(2, 1, 0, 0)
-        >>> k3 = mapdl.k(3, 1, 1, 1)
-        >>> a1 = mapdl.a(k1, k2, k3)
-        >>> a1
-        1
-
-        We can use ``Query.asel`` to interrogate the selection status
-        of the line. The response is an ``enum.IntEnum`` object. If
-        you query a line that does not exist, it will return a status
-        ``SelectionStatus.UNDEFINED``.
-
-        >>> q = Query(mapdl)
-        >>> q.asel(a1)
-        <SelectionStatus.SELECTED: 1>
-        >>> mapdl.asel('NONE')
-        >>> q.asel(a1)
-        <SelectionStatus.UNSELECTED: -1>
-        >>> q.asel(0)
-        <SelectionStatus.UNDEFINED: 0>
-        """
-        response = self._mapdl.run(f'_=ASEL({a})')
-        integer = self._parse_parameter_integer_response(response)
-        return SelectionStatus(integer)
-
-    def esel(self, e: int) -> SelectionStatus:
-        """Returns selection status of an element.
-
-        Returns a ``SelectionStatus`` object with values:
-
-        1  - SELECTED
-        0  - UNDEFINED
-        -1 - UNSELECTED
-
-        Parameters
-        ----------
-        e : int
-            Element number
-
-        Returns
-        -------
-        mapdl.ansys.core.inline_functions.SelectionStatus
-            Status of element
-
-        Examples
-        --------
-        Here we create a single element and interrogate its selection
-        status.
-
-        >>> from ansys.mapdl.core import launch_mapdl
-        >>> from ansys.mapdl.core.inline_functions import Query
-        >>> mapdl = launch_mapdl()
-        >>> mapdl.prep7()
-        >>> mapdl.et(1, 'SHELL181')
-        >>> n1 = mapdl.n(1, 0, 0, 0)
-        >>> n2 = mapdl.n(2, 1, 0, 0)
-        >>> n3 = mapdl.n(3, 1, 1, 1)
-        >>> e1 = mapdl.e(n1, n2, n3)
-        >>> e1
-        1
-
-        We can use ``Query.esel`` to interrogate the selection status
-        of the element. The response is an ``enum.IntEnum`` object. If
-        you query an element that does not exist, it will return a
-        status ``SelectionStatus.UNDEFINED``.
-
-        >>> q = Query(mapdl)
-        >>> q.esel(e1)
-        <SelectionStatus.SELECTED: 1>
-        >>> mapdl.esel('NONE')
-        >>> q.esel(e1)
-        <SelectionStatus.UNSELECTED: -1>
-        >>> q.esel(0)
-        <SelectionStatus.UNDEFINED: 0>
-        """
-        response = self._mapdl.run(f'_=ESEL({e})')
-        integer = self._parse_parameter_integer_response(response)
-        return SelectionStatus(integer)
-
-    def vsel(self, v: int) -> SelectionStatus:
-        """Returns selection status of a volume.
-
-        Returns a ``SelectionStatus`` object with values:
-
-        1  - SELECTED
-        0  - UNDEFINED
-        -1 - UNSELECTED
-
-        Parameters
-        ----------
-        v : int
-            Volume number
-
-        Returns
-        -------
-        mapdl.ansys.core.inline_functions.SelectionStatus
-            Status of element
-
-        Examples
-        --------
-        Here we create a single volume and interrogate its selection
-        status.
-
-        >>> from ansys.mapdl.core import launch_mapdl
-        >>> from ansys.mapdl.core.inline_functions import Query
-        >>> mapdl = launch_mapdl()
-        >>> mapdl.prep7()
-        >>> mapdl.et(1, 'SHELL181')
-        >>> k1 = mapdl.k(1, 0, 0, 0)
-        >>> k2 = mapdl.k(2, 1, 0, 0)
-        >>> k3 = mapdl.k(3, 0, 1, 0)
-        >>> k4 = mapdl.k(4, 0, 0, 1)
-        >>> v1 = mapdl.v(k1, k2, k3, k4)
-        >>> v1
-        1
-
-        We can use ``Query.vsel`` to interrogate the selection status
-        of the element. The response is an ``enum.IntEnum`` object. If
-        you query a volume that does not exist, it will return a
-        status ``SelectionStatus.UNDEFINED``.
-
-        >>> q = Query(mapdl)
-        >>> q.vsel(v1)
-        <SelectionStatus.SELECTED: 1>
-        >>> mapdl.vsel('NONE')
-        >>> q.vsel(v1)
-        <SelectionStatus.UNSELECTED: -1>
-        >>> q.vsel(0)
-        <SelectionStatus.UNDEFINED: 0>
-        """
-        response = self._mapdl.run(f'_=VSEL({v})')
-        integer = self._parse_parameter_integer_response(response)
-        return SelectionStatus(integer)
-
-
-class _LineFractionCoordinateQueries(_ParameterParsing):
-    _mapdl = None
-
-    def lx(self, n: int, lfrac: float) -> float:
-        """X-coordinate of line ``n`` at length fraction ``lfrac``.
-
-        Fetches X-coordinate of line ``n`` at ``lfrac`` times the line
-        length along the line. ``0. <= lfrac <= 1.``.
-
-        Parameters
-        ----------
-        n : int
-            The line number of the line to be considered.
-
-        lfrac: float
-            The fraction of the length of the line along which to
-            enquire. ``0. <= lfrac <= 1.``
-
-        Returns
-        -------
-        float
-            The X-coordinate.
-
-        Examples
-        --------
-        Here we construct a line between the coordinates ``(0, 0, 0)``
-        and ``(1, 2, 3)`` then find the X-coordinate halfway along the line.
-
-        >>> from ansys.mapdl.core.inline_functions import Query
-        >>> from ansys.mapdl.core import launch_mapdl
-        >>> mapdl = launch_mapdl()
-        >>> mapdl.prep7()
-        >>> k0 = mapdl.k(1, 0, 0, 0)
-        >>> k1 = mapdl.k(2, 1, 2, 3)
-        >>> l0 = mapdl.l(k0, k1)
-        >>> q = Query(mapdl)
-        >>> q.lx(l0, 0.5)
-        0.5
-        """
-        response = self._mapdl.run(f'_=LX({n},{lfrac})')
-        return self._parse_parameter_float_response(response)
-
-    def ly(self, n: int, lfrac: float) -> float:
-        """Y-coordinate of line ``n`` at length fraction ``lfrac``.
-
-        Fetches Y-coordinate of line ``n`` at ``lfrac`` times the line
-        length along the line. ``0. <= lfrac <= 1.``.
-
-        Parameters
-        ----------
-        n : int
-            The line number of the line to be considered.
-
-        lfrac: float
-            The fraction of the length of the line along which to
-            enquire. ``0. <= lfrac <= 1.``
-
-        Returns
-        -------
-        float
-            The Y-coordinate.
-
-        Examples
-        --------
-        Here we construct a line between the coordinates ``(0, 0, 0)``
-        and ``(1, 2, 3)`` then find the Y-coordinate halfway along the line.
-
-        >>> from ansys.mapdl.core.inline_functions import Query
-        >>> from ansys.mapdl.core import launch_mapdl
-        >>> mapdl = launch_mapdl()
-        >>> mapdl.prep7()
-        >>> k0 = mapdl.k(1, 0, 0, 0)
-        >>> k1 = mapdl.k(2, 1, 2, 3)
-        >>> l0 = mapdl.l(k0, k1)
-        >>> q = Query(mapdl)
-        >>> q.ly(l0, 0.5)
-        1.0
-        """
-        response = self._mapdl.run(f'_=LY({n},{lfrac})')
-        return self._parse_parameter_float_response(response)
-
-    def lz(self, n: int, lfrac: float) -> float:
-        """Z-coordinate of line ``n`` at length fraction ``lfrac``.
-
-        Fetches Z-coordinate of line ``n`` at ``lfrac`` times the line
-        length along the line. ``0. <= lfrac <= 1.``.
-
-        Parameters
-        ----------
-        n : int
-            The line number of the line to be considered.
-
-        lfrac: float
-            The fraction of the length of the line along which to
-            enquire. ``0. <= lfrac <= 1.``
-
-        Returns
-        -------
-        float
-            The Z-coordinate.
-
-        Examples
-        --------
-        Here we construct a line between the coordinates ``(0, 0, 0)``
-        and ``(1, 2, 3)`` then find the Z-coordinate halfway along the line.
-
-        >>> from ansys.mapdl.core.inline_functions import Query
-        >>> from ansys.mapdl.core import launch_mapdl
-        >>> mapdl = launch_mapdl()
-        >>> mapdl.prep7()
-        >>> k0 = mapdl.k(1, 0, 0, 0)
-        >>> k1 = mapdl.k(2, 1, 2, 3)
-        >>> l0 = mapdl.l(k0, k1)
-        >>> q = Query(mapdl)
-        >>> q.lz(l0, 0.5)
-        1.5
-        """
-        response = self._mapdl.run(f'_=LZ({n},{lfrac})')
-        return self._parse_parameter_float_response(response)
-
-
-class Query(_ComponentQueries,
-            _InverseGetComponentQueries,
-            _DisplacementComponentQueries,
-            _SelectionStatusQueries,
-            _LineFractionCoordinateQueries):
-    """Class containing all the inline functions of APDL.
-
-    Most of the results of these methods are shortcuts for specific
-    combinations of arguments supplied to `mapdl.get(...)` and the
-    same results can be achieved just using `mapdl.get()`.
-
-    Currently implemented functions:
-
-    - `centrx(e)` - get the centroid x-coordinate of element `e`
-    - `centry(e)` - get the centroid y-coordinate of element `e`
-    - `centrz(e)` - get the centroid z-coordinate of element `e`
-    - `nx(n)` - get the x-coordinate of node `n`
-    - `ny(n)` - get the y-coordinate of node `n`
-    - `nz(n)` - get the z-coordinate of node `n`
-    - `kx(k)` - get the x-coordinate of keypoint `k`
-    - `ky(k)` - get the y-coordinate of keypoint `k`
-    - `kz(k)` - get the z-coordinate of keypoint `k`
-    - `ux(n)` - get the structural displacement at node `n` in x
-    - `uy(n)` - get the structural displacement at node `n` in y
-    - `uz(n)` - get the structural displacement at node `n` in z
-    - `node(x, y, z)` - get the node closest to coordinate (x, y, z)
-    - `kp(x, y, z)` - get the keypoint closest to coordinate (x, y, z)
-
-    Examples
-    --------
-    In this example we construct a solid box and mesh it. Then we use
-    the ``Query`` methods ``nx``, ``ny``, and ``nz`` to find the
-    cartesian coordinates of the first node.
-
-    >>> from ansys.mapdl.core import launch_mapdl
-    >>> from ansys.mapdl.core.inline_functions import Query
-    >>> mapdl = launch_mapdl()
-    >>> mapdl.prep7()
-    >>> mapdl.et(1, 'SOLID5')
-    >>> mapdl.block(0, 10, 0, 20, 0, 30)
-    >>> mapdl.esize(2)
-    >>> mapdl.vmesh('ALL')
-    >>> q = Query(mapdl)
-    >>> q.nx(1), q.ny(1), q.nz(1)
-    0.0 20.0 0.0
-    """
-    def __init__(self, mapdl):
-        self._mapdl = mapdl
